add crud to digitcard

+++ src/main/java/com/techgeeknext/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.techgeeknext.repository;
 import com.techgeeknext.model.UserDao;
 import org.springframework.data.repository.CrudRepository;
-public interface UserRepository extends CrudRepository<UserDao, Integer> {
+public interface UserRepository extends CrudRepository<UserDao, Long> {
     UserDao findByUsername(String username);
 }
